fix(version): fix version release workflow
