+++ .changeset/nine-moons-reply.md
@@ -1,0 +1,5 @@
+---
+"clipboard-manager": patch
+---
+
+Fix caching problem with tanstack query
