+++ package.json
@@ -10,8 +10,7 @@
     "preview": "rsbuild preview",
     "tauri": "tauri",
     "lint": "biome check ./src",
-    "semantic-release": "semantic-release",
-    "semantic-release:dry": "semantic-release --dry-run"
+    "version-packages": "changeset version && node scripts/sync-version.js"
   },
   "dependencies": {
     "@tanstack/react-query": "^5.75.7",
@@ -34,23 +33,19 @@
     "@babel/core": "^7.27.1",
     "@babel/plugin-syntax-jsx": "^7.27.1",
     "@biomejs/biome": "^1.9.4",
+    "@changesets/changelog-github": "^0.5.1",
+    "@changesets/cli": "^2.29.3",
     "@emotion/babel-plugin": "^11.13.5",
     "@emotion/react": "^11.14.0",
     "@iarna/toml": "^2.2.5",
     "@rsbuild/core": "^1.3.18",
     "@rsbuild/plugin-react": "^1.3.1",
-    "@semantic-release/changelog": "^6.0.3",
-    "@semantic-release/exec": "^7.1.0",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^11.0.3",
     "@swc/plugin-emotion": "^9.0.4",
     "@tauri-apps/cli": "^2",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "babel-loader": "^10.0.0",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
-    "conventional-changelog-conventionalcommits": "^9.0.0",
-    "semantic-release": "^24.2.5",
     "typescript": "~5.6.2"
   }
 }
